% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcarsize.R
\name{mtcarsize}
\alias{mtcarsize}
\title{Anonomize labels and text in a ggplot2 object to make it look like mtcars}
\usage{
mtcarsize(x)
}
\arguments{
\item{x}{a ggplot object}
}
\value{
a ggplot object
}
\description{
Anonomize labels and text in a ggplot2 object to make it look like mtcars
}
\examples{
library(ggplot2)
fff <- ggplot(mtcars, aes(mpg, disp, color = factor(vs), shape = factor(am))) +
  geom_point()

fff

mtcarsize(fff)
}
